refactor: valid 책임을 Generator -> Lotto 객체로 변경
